requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
chartjs==1.2
Django==4.1.7
docopt==0.6.2
idna==3.4
imutils==0.5.4
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
pipreqs==0.4.11
pyocr==0.8.3
requests==2.28.2
scipy==1.10.1
sqlparse==0.4.3
urllib3==1.26.15
yarg==0.1.9
